move control panel into separate component
